Mover view del boton al view de los videos
